Optimize codegen/parser size
Reduce cc_codegen from 30268B to 26527B by simplifying codegen paths, consolidating tables/strings, and trimming unused fields. Align parser counts and tighten lexer/parser loops for ZOS parity.

+++ src/common/symbol_table.c
@@ -15,7 +15,7 @@ symbol_table_t* symbol_table_create(symbol_table_t* parent) {
         return NULL;
     }
 
-    for (size_t i = 0; i < SYMBOL_TABLE_SIZE; i++) {
+    for (uint8_t i = 0; i < SYMBOL_TABLE_SIZE; i++) {
         table->buckets[i] = NULL;
     }
